+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Fix supported Ledger companion app
 - Fix 2FA reactivation on liquid pre-csv utxos
+- Fix validation on bip21 liquid uri
 
 ## [5.0.6] - 2025-07-02